+++ example/lib/zillow/box.dart
@@ -100,7 +100,7 @@ class _BoxPageState extends State<BoxPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: const Text('SelectorBox')),
+      appBar: AppBar(title: const Text('SelectBox')),
       body: SafeArea(
         child: Scrollbar(
           child: SingleChildScrollView(
@@ -113,7 +113,7 @@ class _BoxPageState extends State<BoxPage> {
                   'Neighborhood',
                   style: TextStyle(fontSize: 20),
                 ),
-                SelectorBox(
+                SelectBox(
                   margin:
                       const EdgeInsets.symmetric(horizontal: 16, vertical: 10),
                   delegate: CascadingSelectorDelegate(
@@ -142,7 +142,7 @@ class _BoxPageState extends State<BoxPage> {
                   'Price',
                   style: TextStyle(fontSize: 20),
                 ),
-                SelectorBox(
+                SelectBox(
                   margin:
                       const EdgeInsets.symmetric(horizontal: 16, vertical: 10),
                   delegate: GridSelectorDelegate(
@@ -172,7 +172,7 @@ class _BoxPageState extends State<BoxPage> {
                   'Price (Range Slider)',
                   style: TextStyle(fontSize: 20),
                 ),
-                SelectorBox(
+                SelectBox(
                   margin:
                       const EdgeInsets.symmetric(horizontal: 16, vertical: 10),
                   delegate: ListSelectorDelegate(
@@ -193,7 +193,7 @@ class _BoxPageState extends State<BoxPage> {
                   'Rooms',
                   style: TextStyle(fontSize: 20),
                 ),
-                SelectorBox(
+                SelectBox(
                   margin:
                       const EdgeInsets.symmetric(horizontal: 16, vertical: 10),
                   delegate: FlattenSelectorDelegate(
@@ -217,7 +217,7 @@ class _BoxPageState extends State<BoxPage> {
                   'Sort',
                   style: TextStyle(fontSize: 20),
                 ),
-                SelectorBox(
+                SelectBox(
                   margin:
                       const EdgeInsets.symmetric(horizontal: 16, vertical: 10),
                   delegate: ListSelectorDelegate(